requirements.txt
google-generativeai>=0.7.0
langchain-google-genai>=1.0.0
langchain>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0
chromadb>=0.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
requests>=2.31.0
